+++ samples/integration/streaming.cpp
@@ -255,9 +255,7 @@ void StartStreaming(unsigned int outputWidth, unsigned int outputHeight, unsigne
 	}
 
 	// Setup the video parameters
-	TTV_VideoParams videoParams;
-	memset(&videoParams, 0, sizeof(videoParams));
-	videoParams.size = sizeof(TTV_VideoParams);
+	TTV_VideoParams videoParams = {sizeof(videoParams)}
 	videoParams.outputWidth = outputWidth;
 	videoParams.outputHeight = outputHeight;
 	videoParams.targetFps = targetFps;